+++ src/main/scala/domain/transform/calculations/equal/NumberEqualCalculationsScala.scala
@@ -4,8 +4,8 @@ import java.util.concurrent.ThreadLocalRandom
 
 import domain.in.field.options.NumberRangeScala
 
-final class NumberEqualCalculationsScala(numberRanges: Seq[NumberRangeScala]) extends EqualCalculationsScala[Float] {
-  override def calculate: Float = {
+final class NumberEqualCalculationsScala(numberRanges: Seq[NumberRangeScala]) extends EqualCalculationsScala[Double] {
+  override def calculate: Double = {
     val numberRange = this.numberRanges(ThreadLocalRandom.current.nextInt(0, this.numberRanges.size))
 
     val fMin = numberRange.min
