+++ src/main/scala/domain/transform/rule/create/CreateRuleNumberScala.scala
@@ -4,8 +4,8 @@ import domain.in.rule.InputRuleScala
 import domain.transform.rule.condition.RulesConditionNumberScala
 import domain.value.ValueScala
 
-final class CreateRuleNumberScala extends CreateRuleTScala[Float, RulesConditionNumberScala] {
-  override def getCondition(output: ValueScala[Float], inputRule: InputRuleScala[Float]): RulesConditionNumberScala = {
+final class CreateRuleNumberScala extends CreateRuleTScala[Double, RulesConditionNumberScala] {
+  override def getCondition(output: ValueScala[Double], inputRule: InputRuleScala[Double]): RulesConditionNumberScala = {
     new RulesConditionNumberScala(output.value, inputRule.condition, inputRule.comparator)
   }
 }
